MAINT: spatial.distance: move files into their own submodule dir
