+++ pyrefly.toml
@@ -50,8 +50,8 @@ ignore-missing-imports = [
     "scipy.sparse._sparsetools",
     "scipy.sparse.linalg._dsolve._superlu",
     "scipy.sparse.linalg._eigen.arpack._arpacklib",
-    "scipy.spatial._distance_pybind",
-    "scipy.spatial._distance_wrap",
+    "scipy.spatial.distance._distance_pybind",
+    "scipy.spatial.distance._distance_wrap",
     "scipy.spatial.cKDTree", # TODO(lucascolley): figure out why this is needed, I can't reproduce locally
     "scipy.special._gufuncs",
     "scipy.special._special_ufuncs",
