+++ scipy/spatial/distance/_distance.pyi
