+++ scipy/spatial/distance/src/distance_impl.h
