+++ scipy/spatial/distance/src/distance_metrics.h
