+++ scipy/spatial/distance/src/function_ref.h
