+++ scipy/spatial/distance/src/views.h
