+++ scipy/spatial/distance/tests/data/cdist-X1.txt
