+++ scipy/spatial/distance/tests/data/cdist-X2.txt
