+++ scipy/spatial/distance/tests/data/iris.txt
