+++ scipy/spatial/distance/tests/data/pdist-boolean-inp.txt
