+++ scipy/spatial/distance/tests/data/pdist-chebyshev-ml-iris.txt
