+++ scipy/spatial/distance/tests/data/pdist-cityblock-ml-iris.txt
