+++ scipy/spatial/distance/tests/data/pdist-cityblock-ml.txt
