+++ scipy/spatial/distance/tests/data/pdist-cosine-ml-iris.txt
