+++ scipy/spatial/distance/tests/data/pdist-double-inp.txt
