+++ scipy/spatial/distance/tests/data/pdist-euclidean-ml.txt
