+++ scipy/spatial/distance/tests/data/pdist-minkowski-3.2-ml-iris.txt
