+++ scipy/spatial/distance/tests/data/pdist-minkowski-5.8-ml-iris.txt
