+++ scipy/spatial/distance/tests/data/pdist-seuclidean-ml-iris.txt
