+++ scipy/spatial/distance/tests/data/pdist-spearman-ml.txt
