+++ scipy/spatial/distance/tests/data/random-bool-data.txt
