+++ scipy/spatial/distance/tests/data/random-double-data.txt
