+++ scipy/spatial/distance/tests/data/random-int-data.txt
