+++ scipy/spatial/distance/tests/data/random-uint-data.txt
