+++ scipy/spatial/distance/tests/test_hausdorff.py
