+++ scipy/spatial/tests/meson.build
@@ -2,8 +2,6 @@ py3.install_sources([
     '__init__.py',
     'test__plotutils.py',
     'test__procrustes.py',
-    'test_distance.py',
-    'test_hausdorff.py',
     'test_kdtree.py',
     'test_qhull.py',
     'test_slerp.py',
@@ -14,36 +12,7 @@ py3.install_sources([
 )
 
 py3.install_sources([
-    'data/cdist-X1.txt',
-    'data/cdist-X2.txt',
     'data/degenerate_pointset.npz',
-    'data/iris.txt',
-    'data/pdist-boolean-inp.txt',
-    'data/pdist-chebyshev-ml-iris.txt',
-    'data/pdist-chebyshev-ml.txt',
-    'data/pdist-cityblock-ml-iris.txt',
-    'data/pdist-cityblock-ml.txt',
-    'data/pdist-correlation-ml-iris.txt',
-    'data/pdist-correlation-ml.txt',
-    'data/pdist-cosine-ml-iris.txt',
-    'data/pdist-cosine-ml.txt',
-    'data/pdist-double-inp.txt',
-    'data/pdist-euclidean-ml-iris.txt',
-    'data/pdist-euclidean-ml.txt',
-    'data/pdist-hamming-ml.txt',
-    'data/pdist-jaccard-ml.txt',
-    'data/pdist-jensenshannon-ml-iris.txt',
-    'data/pdist-jensenshannon-ml.txt',
-    'data/pdist-minkowski-3.2-ml-iris.txt',
-    'data/pdist-minkowski-3.2-ml.txt',
-    'data/pdist-minkowski-5.8-ml-iris.txt',
-    'data/pdist-seuclidean-ml-iris.txt',
-    'data/pdist-seuclidean-ml.txt',
-    'data/pdist-spearman-ml.txt',
-    'data/random-bool-data.txt',
-    'data/random-double-data.txt',
-    'data/random-int-data.txt',
-    'data/random-uint-data.txt',
     'data/selfdual-4d-polytope.txt'
   ],
   subdir: 'scipy/spatial/tests/data',
